Move MultiTrajectoryHelpers to core
